# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TokenOS Runtime",
    "version": "0.2.0"
  },
  "paths": {
...
    "/v1/runs": {
      "post": {
        "summary": "Create Run",
        "operationId": "create_run_v1_runs_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunSpec-Input"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "summary": "Read Run",
        "operationId": "read_run_v1_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Run"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/runtime": {
      "get": {
        "summary": "Info",
        "operationId": "info_v1_runtime_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Info V1 Runtime Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks": {
      "post": {
        "summary": "Run",
        "operationId": "run_v1_tasks_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRecord"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Tasks",
        "operationId": "tasks_v1_tasks_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Tasks V1 Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}": {
      "get": {
        "summary": "Task",
        "operationId": "task_v1_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Task V1 Tasks  Task Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tasks/{task_id}/artifacts/{key}": {
      "get": {
        "summary": "Artifact",
        "operationId": "artifact_v1_tasks__task_id__artifacts__key__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Task Id"
            }
          },
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Key"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Artifact V1 Tasks  Task Id  Artifacts  Key  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents": {
      "post": {
        "summary": "Ingest",
        "operationId": "ingest_v1_documents_post",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Document"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ingest V1 Documents Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Documents",
        "operationId": "documents_v1_documents_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  },
                  "title": "Response Documents V1 Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/tools": {
      "get": {
        "summary": "Tools",
        "operationId": "tools_v1_tools_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Tools V1 Tools Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/mcp/{server}/discover": {
      "post": {
        "summary": "Discover",
        "operationId": "discover_v1_mcp__server__discover_post",
        "parameters": [
          {
            "name": "server",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Server"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Discover V1 Mcp  Server  Discover Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/benchmarks": {
      "get": {
        "summary": "Benchmarks",
        "operationId": "benchmarks_v1_benchmarks_get",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Benchmarks V1 Benchmarks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Ablations": {
        "properties": {
          "full_memory": {
            "type": "boolean",
            "title": "Full Memory",
            "default": false
          },
          "all_tools": {
            "type": "boolean",
            "title": "All Tools",
            "default": false
          },
          "fixed_rag": {
            "type": "boolean",
            "title": "Fixed Rag",
            "default": false
          },
          "full_trajectory": {
            "type": "boolean",
            "title": "Full Trajectory",
            "default": false
          },
          "no_cache": {
            "type": "boolean",
            "title": "No Cache",
            "default": false
          },
          "fixed_workflow": {
            "type": "boolean",
            "title": "Fixed Workflow",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Ablations"
      },
...
      "Document": {
        "properties": {
          "document_id": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_.-]{1,128}$",
            "title": "Document Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "title",
          "text"
        ],
        "title": "Document"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Observation": {
        "properties": {
          "step": {
            "type": "integer",
            "title": "Step"
          },
          "tool": {
            "type": "string",
            "title": "Tool"
          },
          "arguments": {
            "additionalProperties": true,
            "type": "object",
            "title": "Arguments"
          },
          "result": {
            "additionalProperties": true,
            "type": "object",
            "title": "Result"
          },
          "artifact_id": {
            "type": "string",
            "title": "Artifact Id"
          },
          "error": {
            "type": "boolean",
            "title": "Error",
            "default": false
          },
          "cache_hit": {
            "type": "boolean",
            "title": "Cache Hit",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "step",
          "tool",
          "arguments",
          "result",
          "artifact_id"
        ],
        "title": "Observation"
      },
...
      "Rates-Input": {
        "properties": {
          "currency": {
            "type": "string",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency"
          },
          "version": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Version"
          },
          "input_per_million": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Input Per Million"
          },
          "cached_input_per_million": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Cached Input Per Million"
          },
          "output_per_million": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Output Per Million"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "currency",
          "version",
          "input_per_million",
          "cached_input_per_million",
          "output_per_million"
        ],
        "title": "Rates",
        "description": "Currency units per million tokens. Frozen into each run for reproducible cost."
      },
...
      "Run": {
        "properties": {
          "spec": {
            "$ref": "#/components/schemas/RunSpec-Output"
          },
          "spent_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Spent Tokens",
            "default": 0
          },
          "reserved_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Reserved Tokens",
            "default": 0
          },
          "input_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Output Tokens",
            "default": 0
          },
          "cached_input_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Cached Input Tokens",
            "default": 0
          },
          "reasoning_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Reasoning Tokens",
            "default": 0
          },
          "cost": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "breached": {
            "type": "boolean",
            "title": "Breached",
            "default": false
          },
          "version": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "minimum": 0.0,
            "title": "Version",
            "default": 0
          },
          "available_tokens": {
            "type": "integer",
            "title": "Available Tokens",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "spec",
          "available_tokens"
        ],
        "title": "Run"
      },
      "RunSpec-Input": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "budget_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Budget Tokens"
          },
          "context_window": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Context Window"
          },
          "model_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Id"
          },
          "policy_version": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Policy Version"
          },
          "rates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Rates-Input"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "budget_tokens",
          "context_window",
          "model_id",
          "policy_version"
        ],
        "title": "RunSpec"
      },
      "RunSpec-Output": {
        "properties": {
          "run_id": {
            "type": "string",
            "format": "uuid",
            "title": "Run Id"
          },
          "budget_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Budget Tokens"
          },
          "context_window": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Context Window"
          },
          "model_id": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Model Id"
          },
          "policy_version": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Policy Version"
          },
          "rates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Rates-Output"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "budget_tokens",
          "context_window",
          "model_id",
          "policy_version"
        ],
        "title": "RunSpec"
      },
...
      "TaskRecord": {
        "properties": {
          "request": {
            "$ref": "#/components/schemas/TaskRequest"
          },
          "status": {
            "type": "string",
            "enum": [
              "running",
              "completed",
              "budget_exhausted",
              "step_limit",
              "failed"
            ],
            "title": "Status"
          },
          "answer": {
            "type": "string",
            "title": "Answer",
            "default": ""
          },
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Evidence Ids"
          },
          "observations": {
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "type": "array",
            "title": "Observations"
          },
          "decisions": {
            "items": {
              "$ref": "#/components/schemas/Selection"
            },
            "type": "array",
            "title": "Decisions"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "source": {
            "type": "string",
            "enum": [
              "simulated",
              "provider"
            ],
            "title": "Source"
          },
          "model_id": {
            "type": "string",
            "title": "Model Id"
          },
          "policy": {
            "$ref": "#/components/schemas/Policy"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "latency_ms": {
            "type": "number",
            "title": "Latency Ms",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request",
          "status",
          "source",
          "model_id",
          "policy"
        ],
        "title": "TaskRecord"
      },
      "TaskRequest": {
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid",
            "title": "Task Id"
          },
          "query": {
            "type": "string",
            "maxLength": 16000,
            "minLength": 1,
            "title": "Query"
          },
          "history": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 200,
            "title": "History"
          },
          "mode": {
            "type": "string",
            "enum": [
              "adaptive",
              "fixed",
              "full"
            ],
            "title": "Mode",
            "default": "adaptive"
          },
          "budget_tokens": {
            "type": "integer",
            "maximum": 9.223372036854776e+18,
            "exclusiveMinimum": 0.0,
            "title": "Budget Tokens",
            "default": 60000
          },
          "ablations": {
            "$ref": "#/components/schemas/Ablations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "TaskRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}